Update script.js
created a new function for the delete button in addition to adding CSS into the script so that users are able to can visually identify what function they're currently using. Also, I corrected the the buttons and made the play button able to start and resume the audio by adding a flag variable.